fix(templates): #65 - Update artifactID setting
Eclipse does not allow the importing of multiple projects with the same
artifactID.  Currently in situations where you are not promted to anter an
artifactID a default one is set for you - this is problematic.  The change
here is to move to a system where the artifactID is set to the user
specified artifactID or if not provided to the appName.  A subsequent
change will be needed to modify the default to undefined.

Signed-off-by: Duane <applebyd@uk.ibm.com>